Replace panics with inline-asm udf instructions
This removes about 256 bytes of rodata (depending on the particular path
that defmt-serial is built from; the panic call stores the current
filename's full path in a &str), plus 50-some bytes of panic_handler
since it's the last panic in the firmware I'm writing.  This lets that
firmware fit in <1k.